rm redundant includes
