+++ tools/server/server-queue.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "server-common.h"
 #include "server-task.h"
 
 #include <chrono>
