+++ tools/server/server-task.cpp
@@ -1,15 +1,11 @@
+#include "server-common.h"
+#include "server-task.h"
+
 #include "common.h"
 #include "llama.h"
 #include "chat.h"
 #include "sampling.h"
 #include "json-schema-to-grammar.h"
-
-#include <string>
-#include <unordered_set>
-#include <list>
-
-#include "server-common.h"
-#include "server-task.h"
 
 using json = nlohmann::ordered_json;
